requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.81.0
packaging==25.0
protobuf==5.29.4
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@fdc6ca50c19f1c85971437a072d89d4e5ce024b8
PyAudio==0.2.14
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.8.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
